[TYPESCRIPT] Ajoute les types communs au LAB dans le fichier dédié commun.ts
